+++ tools/package/package.ps1
@@ -18,7 +18,7 @@
 #>
 [CmdletBinding()]
 param(
-    [string]$Version = "0.2.0"
+    [string]$Version = "0.2.1"
 )
 
 $ErrorActionPreference = 'Stop'